Remove ITR_NEW instruction code
