+++ bvm/ballerina-core/src/main/java/org/ballerinalang/util/codegen/Mnemonics.java
@@ -277,7 +277,6 @@ public class Mnemonics {
 
         mnemonics[InstructionCodes.CONVERT] = "convert";
 
-        mnemonics[InstructionCodes.ITR_NEW] = "itr_new";
         mnemonics[InstructionCodes.ITR_NEXT] = "itr_next";
         mnemonics[InstructionCodes.INT_RANGE] = "int_range";
